+++ browser-tests/utils/settings.ts
@@ -2,52 +2,44 @@
 require('dotenv').config({ path: '.env' });
 
 const LOCAL_URL = 'http://localhost:3000';
-const TEST_URL = 'https://jassari.test.kuva.hel.ninja';
 
 export const username = (): string => {
-  if (!process.env.REACT_APP_TESTING_USERNAME) {
-    throw new Error('No REACT_APP_TESTING_USERNAME specified');
+  if (!process.env.BROWSER_TESTING_USERNAME) {
+    throw new Error('No BROWSER_TESTING_USERNAME specified');
   }
-  return process.env.REACT_APP_TESTING_USERNAME;
+  return process.env.BROWSER_TESTING_USERNAME;
 };
 
 export const usernameWithExistingProfile = (): string => {
-  if (!process.env.REACT_APP_TESTING_USERNAME_WITH_PROFILE) {
-    throw new Error('No REACT_APP_TESTING_USERNAME specified');
+  if (!process.env.BROWSER_TESTING_USERNAME_WITH_PROFILE) {
+    throw new Error('No BROWSER_TESTING_USERNAME specified');
   }
-  return process.env.REACT_APP_TESTING_USERNAME_WITH_PROFILE;
+  return process.env.BROWSER_TESTING_USERNAME_WITH_PROFILE;
 };
 
 export const password = (): string => {
-  if (!process.env.REACT_APP_TESTING_PASSWORD) {
-    throw new Error('No REACT_APP_TESTING_PASSWORD specified');
+  if (!process.env.BROWSER_TESTING_PASSWORD) {
+    throw new Error('No BROWSER_TESTING_PASSWORD specified');
   }
 
-  return process.env.REACT_APP_TESTING_PASSWORD;
+  return process.env.BROWSER_TESTING_PASSWORD;
 };
 
 export const testURL = () => {
-  switch (process.env.REACT_APP_ENVIRONMENT) {
-    case 'local':
-      return LOCAL_URL;
-    case 'staging':
-      return TEST_URL;
-    default:
-      return LOCAL_URL;
-  }
+  return process.env.BROWSER_TESTING_URL || LOCAL_URL;
 };
 
 export const mailosaurApiKey = (): string => {
-  if (!process.env.REACT_APP_MAILOSAUR_API) {
-    throw new Error('No REACT_APP_MAILOSAUR_API specified');
+  if (!process.env.BROWSER_TESTING_MAILOSAUR_API) {
+    throw new Error('No BROWSER_TESTING_MAILOSAUR_API specified');
   }
 
-  return process.env.REACT_APP_MAILOSAUR_API;
+  return process.env.BROWSER_TESTING_MAILOSAUR_API;
 };
 export const mailosaurServerId = (): string => {
-  if (!process.env.REACT_APP_MAILOSAUR_SERVER_ID) {
-    throw new Error('No REACT_APP_MAILOSAUR_SERVER_ID specified');
+  if (!process.env.BROWSER_TESTING_MAILOSAUR_SERVER_ID) {
+    throw new Error('No BROWSER_TESTING_MAILOSAUR_SERVER_ID specified');
   }
 
-  return process.env.REACT_APP_MAILOSAUR_SERVER_ID;
+  return process.env.BROWSER_TESTING_MAILOSAUR_SERVER_ID;
 };